requirements.txt
requests~=2.21.0
dhooks~=1.1.4
pyyaml~=5.4.1